changed int types to work with cosmosdb

+++ internal/team/GetTeam_test.go
@@ -21,9 +21,9 @@ func TestGetTeamById(t *testing.T) {
 				bson.D{
 					{Key: "_id", Value: id},
 					{Key: "name", Value: "test"},
-					{Key: "owner", Value: int64(1)},
+					{Key: "owner", Value: int32(1)},
 					{Key: "membersWithoutOwner", Value: bson.A{int64(1)}},
-					{Key: "score", Value: int64(0)},
+					{Key: "score", Value: int32(0)},
 					{Key: "data", Value: map[string]string{"test": "test"}},
 				},
 			}, nil, nil)
@@ -58,9 +58,9 @@ func TestGetTeamByIdName(t *testing.T) {
 				bson.D{
 					{Key: "_id", Value: id},
 					{Key: "name", Value: teamName},
-					{Key: "owner", Value: int64(1)},
+					{Key: "owner", Value: int32(1)},
 					{Key: "membersWithoutOwner", Value: bson.A{int64(1)}},
-					{Key: "score", Value: int64(0)},
+					{Key: "score", Value: int32(0)},
 					{Key: "data", Value: map[string]string{"test": "test"}},
 				},
 			}, nil, nil)
